requirements.txt
playwright==1.42.0
beautifulsoup4==4.12.3
gspread==6.0.2
google-api-python-client==2.122.0
google-auth==2.28.2
Pillow==10.2.0
